structure: the cached pan/volume are LONG (what DirectSound writes), spelt 'long'
GetPan/GetVolume take LONG* and all four call sites cast an i32's address. Declaring
them long (the builtin, so DirectSoundMgr.h stays Win32-free as it was written to be)
removes the casts and matches what the API actually writes.

Also annotated: the WAVEFORMAT dword move (retail moves the two adjacent SDK u16
fields as ONE dword and the struct offers no combined member), the RIFF wire walk,
and the remaining zDArray byte-pool accessors in TypeKeyColl.

+++ include/Dsndmgr/DirectSoundMgr.h
@@ -70,8 +70,9 @@ public:
     SoundDevice* m_owner;         // +0x10  owning device back-pointer
     u32 m_playFlags;              // +0x14  Play/looping flags (bit 0 = loop)
     u32 m_freq;                   // +0x18  cached frequency (GetFrequency)
-    i32 m_pan;                    // +0x1c  cached pan (GetPan)
-    i32 m_volume;                 // +0x20  cached volume (GetVolume)
+    long m_pan;                   // +0x1c  cached pan (LONG, the type GetPan writes;
+                                  //        spelt `long` so this header stays Win32-free)
+    long m_volume;                // +0x20  cached volume (LONG, GetVolume writes it)
     u32 m_setFreq;                // +0x24  cached SetFrequency value
     u32 m_durationMs;             // +0x28  duration (ComputeDuration)
     u32 m_sampleCount;            // +0x2c  sample count (set by the clone ctor)
